requirements.txt
black==24.4.2
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
easyocr==1.7.1
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
identify==2.6.13
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.1
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==3.8.0
pyclipper==1.3.0.post6
pyparsing==3.2.3
pytest==8.3.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ruff==0.5.7
scikit-image==0.25.2
scipy==1.16.1
shapely==2.1.1
six==1.17.0
sympy==1.14.0
tifffile==2025.6.11
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.34.0
torch
torchvision
torchaudio
easyocr
pymupdf
pillow
opencv-python-headless
pytest
invoke
python-dateutil
streamlit
sqlite3
cv2
